refactor : remove logs, change the location of notification handler
