+++ src/shared/ui/Dock.tsx
@@ -16,7 +16,7 @@ export default function Dock() {
   const current = info[info.length - 1].name as DockItem
 
   return (
-    <div className='dock box-shadow-dock container-mobile fixed right-0 bottom-0 left-0 z-60 flex h-20 items-center justify-between p-7'>
+    <div className='dock box-shadow-dock container-mobile fixed right-0 bottom-0 left-0 z-60 flex h-18 items-center justify-between border-none p-7'>
       {DOCK_ITEMS.map(item => (
         <DockButton key={item} item={item} selected={current === item} />
       ))}
@@ -27,7 +27,7 @@ export default function Dock() {
 const DockButton = ({ item, selected }: DockButtonProps) => {
   const { replace } = useFlow()
   const onClick = () => {
-    replace(item, { animate: false })
+    replace(item, { animate: false }, { animate: false })
   }
 
   return (
